Phase 2 (1.b): relocate drawing code from LayoutManager into the panes
Each pane now owns its drawing. The per-pane _draw_* method clusters move
verbatim from LayoutManager into their pane classes (DatePane, SubwayPane,
HourlyWeatherPane, CitibikePane, WeatherOverviewPane); the clusters were already
disjoint, so nothing is shared across panes. The Pane base carries the same
config shortcuts LayoutManager had (self.display/weather/subway/time), which let
the bodies move unchanged. The deferred ui.layout imports from 1.a are gone.

LayoutManager is now just the chrome holder (_draw_sections), still wired into
the Screen as chrome_fn. test_no_trains_message points at SubwayPane, since
_no_trains_message moved there.

Behavior-preserving: 24 golden views at zero pixel diff; 32 tests pass.
